// types.ts
export interface Staff {
  id: number;
  fullName: string;
  role: string;
  department: string;
  status: 'Active' | 'On Leave';
  email: string;
  phone: string;
  salary: number;
}

export interface LeaveRequest {
  id: number;
  staffId: number;
  type: 'Sick' | 'Annual' | 'Personal';
  startDate: string;
  endDate: string;
  reason: string;
  status: 'Pending' | 'Approved' | 'Rejected';
}